Line-based drawing initial code
This will be the way the screen will be drawn, as it is the
way the emulator flow behaves. The tile-based approach was too stiff
against some requirements such as changing the SC and WD registers
during H-BLANK.
Line-based drawing allows this changes easily, also the partial window
drawing and it also allows the step-sized drawing!
I call that a win-win-win.