Separated drawing functions into a separate static class
Now they're merely functions that are used to modify a bitmap
according to a DisplayDefinition. This permits to separate the
actual display object from the drawing functionality.

This will enable easier construction of debug tools and displays
by not having to "mock" the display state to get the correct output.