Code reordering preparing for a incremental approach
Right now each step is drawed separatedly, but some steps require information
from previous steps. Instead of codifying that information in some way in the pixel
and using it at the image merge, the codification will be used when needed by
writing directly into the final target incrementally (the way the actual GB works).

The other targets will still be created, but will not be used for any purpose
other than extra-information.