Drawing now done in pipeline
Before, each step (background, window, sprites) were made into
a separate target (frame buffer) and then combined (this is still done,
but for debug purposes).

Now, as some steps (particularly sprites) need information on what is
on the screen PER sprite, we need to draw the sprites directly into
the same target in order to have a convenient place to react to the present pixels.

As a plus, the code is much more concise and better to read.
Win-Win-Win