Hooked the DMA & memory to the Gameboy step loop
Now the the DMA will get called with the amount of ticks for each step.
Note that the DMA is a memory's attribute and the memory handler only has a
reference to it. I think this make more sense, because otherwise the MemoryHandler
would have to have a Step method just to pass that update to the DMA.

It is bad enought that the Memory has to have such method only to update the DMA
(though semantically is not that bad, because the DMA *is* a memory feature).