Added operands information to the CPU View
Now, foreach instruction, we can see what the literals are.
This will enable easier visualization (and debugging!) of what's
happening step-by-step. Maybe a literal interpretation could be
useful... maybe.

BTW, apparently there is a bug with the current PC being displayed.
The operands seems to be off-by-one with the current PC instruction
displayed.