Moved the instruction lambdas to the bottom of CPU
It was very annoying having to scroll 2000 lines to find
the rest of the CPU definitions.

Would be nice to see if C# has some kind of C-line include,
which is a merely copy of the text before the compiler starts.
With this, we could separate the file without having to change
the lambdas.