Now the memory updates when a step writes into memory
This only happens through a "high-level" write, not a CPU
internal function write.
It works through a new event, that changes what the memory should
highlight. It is job of the stepHandler to update this though
(otherwise we would be updating for the view for every write, which
is just too much)

TODO: That the view scrolls correctly to the highlighter value