FIX: Interrupts woudl wrongly add to nextPC
Interrupts are implemented with a CALL (because the PC is inserted into the stack).
Technically interrupts are not a call, but a insert into stack +
a jump into an address, but that is not the problem.
The problem is that the CPU assumed that a CALL instruction passed, thus
adding the CALL length into the nextPC. This is wrong: the address to
be pushed into the stack is the PC of the address *before* the interruption.