WIP: On-demand disassembly
Currently the disassembly is static following conditional jumps.
It should not follow conditional jumps as there is no guarantee that
the code is structured in a way that is assured that a particular branch
will be followed.
Instead it should only disassemble the instruction are sure to be runned
from the entry point. Later, if a unknown address is found, it should
be disassembled from that point and acumulate the results.
This way we only disassemble the code we strictly need and are sure
to be aligned with the program