Completed disassembly permissive mode
This mode is the same as the disassembler we had before:
assumes conditional jumps and CALLs should be followed by
both branches. Non-permissive mode doesn't make an assumption
because there is no guarantee that a program will ever take
a particular branch, so there is no guarantee that a particular
branch are instructions.