WriteableBitmap experiment for drawing speed
This will enable to write memory *directly* to the image back-buffer
instead of the strange copying memory we're using now.

But the main win in speed is that the gameboy display can process the
data in plain arrays instead of Bitmaps, where the lock/unlock process
is just to expensive.
This should give us all the speed we need.
This "experiment" proves it can be done. The final code should require
very little changes in the display code.
Good times.