Added separete update foreach debug target
This will enable to only update (both in GB and in the view) the
debug targets that are required. This enables to start with only
the minimum work needed for emulating and then extra optional
debug targets can be added.

The promise of 60 FPS should be mantained *nominally* only for
the Screen alone without any additional debug targets.