Ramblings on timing loops
Trying to tie the loop to the VBlank event and *that* to 60 FPS
instead of synchronizing the loops into an arbitrary 5 ms boundary.

Works, though it doesn't work much better than the other loop.
Design-wise speaking, permites to sync both video and sound to a
constant game loop, which could proove very useful.

As for now, the old loop is running.