Separated timing from actual display enabled
This means that the display will be triggering at 60 FPS
even if the display is off. The actual trigger will now be at
end of VBlank instead of VBlank begin. It's more correct as we
can call a endframe call from gameboy and it will be correct at
display level.