OAM loading only occurs when that memory changes
Before it happened once a line. This itself is not too much a win, because
it's a simple memory copy of 160 bytes. But what will be bigger is the sorting
win of the sprites. As we know when they change, we can sort them only when needed
instead of several times a frame.