Zero delay sound output
Fucking Finally...
The solution was to invert the functionality. Instead to precompute
the amount of sound according to what time the game loop guessed it was passed,
we generate the sound on demand. As the sound card ask for samples we generate them.
This ensures the minimum amount of delay as no single extra sample is generated.

And, as long the sound code is fast enough, we should never get an audio break.
This is clearly the better solution