Interrupts now follow a more "event-driven" model
Before, for every step, the CPU checked the memory values to see
if an interrupt was required. Now when either the IF or IE are written
to, the CPU is flagged to check for interrupts. This reduces the
amount of checks for interrupts a lot.
The only interesting bit is that after an interrupt is found, the CPU
has to immediatelly check for interrupts to see if another should be chained.