Back to a CPU-timing-based sample generation
Instead of sending events, the step will generate the samples directly.
(The events send the same amount in the end, though much harder and
error prone to decode). This is a good solution between the previous
strategy, as the sample generation code is merely moved to the step
procedure.

Now it's a problem of correctly synchronizing the two threads and
dealing with error cases.