New sound generation scheme (hopefully final)
This scheme samples the state of the APU when a sample generation
call is made and sound is generated according those parameters.
Though this will introduce some aliasing over the state of the APU (and
the actual correct samples generated),  this effect is negligible as the
sample generation code is called at a much bigger frequency that any
significant change can occur over the sound hardware. This means that
the output is quite faithful to the actual hardware emulation.

So now we have the objective of having the APU being simulated according
the CPU simulation and outputting sound accordingly.