Changed length simulator to correct counter in sq channel
Before I was merely creating a counter to how much that counter
should last time-wise. Now I treat the value as it actually works in
the GB: it's a counter that clocks at a much lower frequency.

Now I create a timer that clocks at the correct frequency and use
the counter as I should. Though conceptwise it's a little more complex
(and very little explained in the docs), it's much simpler to maintain.