Sound Event Queue didn't account for on hold time
Each event has how much amount has passed since the previous event.
The consumer of the queue didn't keep track of how much time it spent
*waiting* for an event.
This would make when an event appeared, we would wait more time because now
we have to spin on the diff of time of the event when we already did!